+++ .github/workflows/test.yml
@@ -10,6 +10,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
+      with:
+        submodules: recursive
     - uses: actions-rs/toolchain@v1
       with:
         profile: minimal
@@ -27,6 +29,8 @@ jobs:
         arch: [x86_64, riscv64, aarch64]
     steps:
     - uses: actions/checkout@v3
+      with:
+        submodules: recursive
     - uses: actions-rs/toolchain@v1
       with:
         profile: minimal
@@ -37,6 +41,8 @@ jobs:
         crate: cargo-binutils
         version: latest
         use-tool-cache: true
+    - name: Install gcc multilib
+      run: sudo apt update && sudo apt install -y gcc-multilib
     - name: Download musl toolchain
       run: |
         if [ "${{ matrix.arch }}" = "x86_64" ]; then
@@ -49,6 +55,7 @@ jobs:
         wget https://musl.cc/$MUSL_PATH.tgz
         tar -xf $MUSL_PATH.tgz
         mv $MUSL_PATH musl
+        echo "$PWD/musl/bin" >> $GITHUB_PATH
 
     - name: Cache QEMU
       id: cache-qemu
@@ -74,6 +81,5 @@ jobs:
 
     - name: Run app tests
       run: |
-        export PATH=$PATH:$PWD/musl/bin
         make disk_img
         make test ARCH=${{ matrix.arch }}
