+++ crates/allocator_test/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "allocator"
+name = "allocator_test"
 version = "0.1.0"
 edition = "2021"
-authors = ["Yuekai Jia <equation618@gmail.com>"]
+authors = ["Yibin Zhang <zhangyb20@mails.tsinghua.edu.cn>"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 description = "Various allocator algorithms in a unified interface"
@@ -17,7 +17,6 @@ bitmap-allocator = { git = "https://github.com/rcore-os/bitmap-allocator.git", r
 slab_allocator = { path = "../slab_allocator" }
 basic_allocator = { path = "../basic_allocator" }
 tlsf_allocator = { path = "../tlsf_allocator" }
-allocator_test = { path = "../allocator_test" }
 log = "0.4"
 
 [build-dependencies]
