+++ crates/tlsf_allocator/src/lib.rs
@@ -1,6 +1,3 @@
-//! TLSF allocator for `no_std` systems.
-//! written by rust code
-
 #![feature(allocator_api)]
 #![no_std]
 
@@ -13,20 +10,19 @@ use core::mem::size_of;
 mod data;
 use data::*;
 
-/// the heap structure of the allocator
-pub struct Heap {
-    head: AddrPointer, // 指向TLSF头结构的指针
-    total_mem: usize,  // 总共占用内存
-    used_mem: usize,   // 已经分配出去的内存
-    avail_mem: usize,  // 实际可用的内存
+pub struct HeapRust {
+    head: AddrPointer,
+    total_mem: usize, // 总共占用内存
+    used_mem: usize,  // 已经分配出去的内存
+    avail_mem: usize, // 实际可用的内存
 }
 
-unsafe impl Send for Heap {}
+unsafe impl Send for HeapRust {}
 
-impl Heap {
+impl HeapRust {
     /// Create an empty heap
     pub const fn new() -> Self {
-        Heap {
+        HeapRust {
             head: AddrPointer { addr: 0 },
             total_mem: 0,
             used_mem: 0,
@@ -152,15 +148,12 @@ impl Heap {
                 if !(nxt_block.is_null()) {
                     nxt_block.set_prev_phy_pointer(new_block);
                 }
-
                 //设置块大小
                 block.set_size(size);
                 new_block.set_size(nsize - size - 2 * size_of::<usize>()); //别忘了减去新块的头部大小
-
-                //设置使用状态
+                                                                           //设置使用状态
                 block.set_used();
                 new_block.set_free();
-
                 //插回到链表中去
                 self.head.add_into_list(new_block);
                 self.avail_mem -= 2 * size_of::<usize>();
@@ -230,18 +223,104 @@ impl Heap {
         self.head.add_into_list(nblock);
     }
 
-    /// get total bytes
+    /// 查询内存使用情况
     pub fn total_bytes(&self) -> usize {
         self.total_mem
     }
 
-    /// get used bytes
     pub fn used_bytes(&self) -> usize {
         self.used_mem
     }
 
-    /// get available bytes
     pub fn available_bytes(&self) -> usize {
         self.avail_mem
+    }
+}
+
+use core::ffi::c_ulonglong;
+#[link(name = "tlsf")]
+extern "C" {
+    pub fn tlsf_create_with_pool(mem: c_ulonglong, bytes: c_ulonglong) -> c_ulonglong;
+    pub fn tlsf_add_pool(tlsf: c_ulonglong, mem: c_ulonglong, bytes: c_ulonglong) -> c_ulonglong;
+
+    pub fn tlsf_malloc(tlsf: c_ulonglong, bytes: c_ulonglong) -> c_ulonglong; //申请一段内存
+    pub fn tlsf_memalign(tlsf: c_ulonglong, align: c_ulonglong, bytes: c_ulonglong) -> c_ulonglong; //申请一段内存，要求对齐到align
+    pub fn tlsf_free(tlsf: c_ulonglong, ptr: c_ulonglong); //回收
+}
+
+pub struct HeapC {
+    inner: Option<c_ulonglong>,
+}
+
+impl HeapC {
+    pub const fn new() -> Self {
+        Self { inner: None }
+    }
+    pub fn inner_mut(&mut self) -> &mut c_ulonglong {
+        self.inner.as_mut().unwrap()
+    }
+
+    pub fn inner(&self) -> &c_ulonglong {
+        self.inner.as_ref().unwrap()
+    }
+
+    pub fn init(&mut self, start: usize, size: usize) {
+        unsafe {
+            self.inner = Some(
+                tlsf_create_with_pool(start as c_ulonglong, size as c_ulonglong) as c_ulonglong,
+            );
+        }
+    }
+
+    pub fn add_memory(&mut self, start: usize, size: usize) {
+        unsafe {
+            tlsf_add_pool(
+                *self.inner() as c_ulonglong,
+                start as c_ulonglong,
+                size as c_ulonglong,
+            );
+        }
+    }
+
+    pub fn allocate(&mut self, size: usize, align_pow2: usize) -> Result<usize, AllocError> {
+        if align_pow2 <= 8 {
+            unsafe {
+                let ptr = tlsf_malloc(*self.inner() as c_ulonglong, size as c_ulonglong) as usize;
+                if ptr == 0 {
+                    return Err(AllocError);
+                }
+                Ok(ptr)
+            }
+        } else {
+            unsafe {
+                let ptr = tlsf_memalign(
+                    *self.inner() as c_ulonglong,
+                    align_pow2 as c_ulonglong,
+                    size as c_ulonglong,
+                ) as usize;
+                if ptr == 0 {
+                    return Err(AllocError);
+                }
+                Ok(ptr)
+            }
+        }
+    }
+
+    pub fn deallocate(&mut self, pos: usize, _size: usize, _align_pow2: usize) {
+        unsafe {
+            tlsf_free(*self.inner() as c_ulonglong, pos as c_ulonglong);
+        }
+    }
+
+    pub fn total_bytes(&self) -> usize {
+        0
+    }
+
+    pub fn used_bytes(&self) -> usize {
+        0
+    }
+
+    pub fn available_bytes(&self) -> usize {
+        0
     }
 }
